#86 make pass 3 of 4 tests (in PropertyTest.java)

+++ src/main/java/com/dj/server/api/properties/entity/Property.java
@@ -16,7 +16,7 @@ public class Property {
 
     @Id
     @Column
-    private Integer propKey;
+    private Long propKey;
 
     @Column(length = 100)
     private String propValue;
@@ -24,7 +24,7 @@ public class Property {
     @Column(length = 70)
     private String propContent;
 
-    public Property(Integer propKey, String propValue, String propContent) {
+    public Property(Long propKey, String propValue, String propContent) {
         this.propKey = propKey;
         this.propValue = propValue;
         this.propContent = propContent;
